modified migrations files
we added some changes on columns types

+++ database/migrations/2022_03_19_162215_create__plannings_table.php
@@ -18,9 +18,9 @@ class CreatePlanningsTable extends Migration
             $table->date('date_seance');
             $table->time('start_time');
             $table->time('end_time');
-            $table->integer('idCour');
-            $table->integer('idSalle');
-            $table->integer('idUser');
+            $table->unsignedBigInteger('idCour');
+            $table->unsignedBigInteger('idSalle');
+            $table->unsignedBigInteger('idUser');
             $table->foreign('idCour')->references('id')->on('_Coures');
             $table->foreign('idSalle')->references('id')->on('_Salles');
             $table->foreign('idUser')->references('id')->on('_Users');
